requirements.txt
requests>=2.28.0
zai-sdk>=0.2.2